Added relaxed pathfinding
When strict mode is off (i.e. 'relaxed' is enabled), pathfinding can (and often will) find a path to a neighbouring tile of the goal, instead of the goal itself. This allows it to path to a tile that is not traversible such as tiles with entities or a Water tile at the edge of a lake.