Player can now tell the citizen to chop down trees
The biggest change in this commit is the addition of the selectTile() method in UI. This provides an interface for different parts of the game to request the player to select a tile. The first use of this is the cut down tree command. The player is prompted to select a tile with a tree, and upon selecting one, a task is added which the citizen can then work on.