Split gamesystems into pre-update and post-update
This fixes a bug where saplings growing into trees will cause a brief period where neither the Sapling nor a Tree are on the tile, because of the way the game/district handles creating and then moving entities. MoveSystem would run, so the tree should be given a valid tile before the UI draws, but because the District was not updated beforehand, the tree entity is not in the vector of entities, so it is not updated until the next game tick. This is what caused the bug.